refactor(search): extract hybrid search pipeline into hybrid-search.ts
Move hybridSearch and its private helpers (toRankedList, embedQuery,
vectorSearch, fileContentVectorSearch, runFileContentFts, tryRerank)
out of search-queries.ts into search/hybrid-search.ts, and the two
hybridSearch describe blocks (plus installQueryPoison, their only
consumer) out of search-index.test.ts into hybrid-search.test.ts.

Pure relocation: every moved block is byte-identical to its source.
search-index.ts binds hybridSearch from the new module directly.
